🐛 (Border) - Remove border radius and border color from preview div when download is started
